refactor(ContextMenu): Abstract out `useRightClick` hook
